# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TAOX API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/processes/": {
            "get": {
                "tags": [
                    "processes"
                ],
                "summary": "Get All Processes",
                "description": "Obtiene todos los procesos de una empresa",
                "operationId": "get_all_processes_processes__get",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProcessResponseDTO"
                                    },
                                    "title": "Response Get All Processes Processes  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "processes"
                ],
                "summary": "Create Process",
                "description": "Crea un nuevo proceso",
                "operationId": "create_process_processes__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProcessCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/processes/{id_proceso}": {
            "get": {
                "tags": [
                    "processes"
                ],
                "summary": "Get Process By Id",
                "description": "Obtiene un proceso por ID",
                "operationId": "get_process_by_id_processes__id_proceso__get",
                "parameters": [
                    {
                        "name": "id_proceso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Proceso"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "processes"
                ],
                "summary": "Update Process",
                "description": "Actualiza un proceso",
                "operationId": "update_process_processes__id_proceso__put",
                "parameters": [
                    {
                        "name": "id_proceso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Proceso"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProcessUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "processes"
                ],
                "summary": "Delete Process",
                "description": "Elimina un proceso",
                "operationId": "delete_process_processes__id_proceso__delete",
                "parameters": [
                    {
                        "name": "id_proceso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Proceso"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProcessCreateDTO": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Nombre",
                        "description": "Nombre del proceso"
                    },
                    "descripcion": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Descripcion",
                        "description": "Descripción del proceso"
                    },
                    "tipo_proceso": {
                        "type": "string",
                        "title": "Tipo Proceso",
                        "description": "Tipo de proceso (produccion, prestamo, almacenamiento, descarte)"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa",
                        "description": "ID de la empresa propietaria"
                    }
                },
                "type": "object",
                "required": [
                    "nombre",
                    "tipo_proceso",
                    "id_empresa"
                ],
                "title": "ProcessCreateDTO",
                "description": "DTO para crear un nuevo proceso",
                "example": {
                    "descripcion": "Área destinada a la producción de químicos",
                    "id_empresa": "company-123",
                    "nombre": "Producción de químicos",
                    "tipo_proceso": "produccion"
                }
            },
            "ProcessResponseDTO": {
                "properties": {
                    "id_proceso": {
                        "type": "integer",
                        "title": "Id Proceso"
                    },
                    "nombre": {
                        "type": "string",
                        "title": "Nombre"
                    },
                    "descripcion": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Descripcion"
                    },
                    "tipo_proceso": {
                        "type": "string",
                        "title": "Tipo Proceso"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id_proceso",
                    "nombre",
                    "descripcion",
                    "tipo_proceso",
                    "id_empresa",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProcessResponseDTO",
                "description": "DTO para respuestas de proceso",
                "example": {
                    "created_at": "2025-03-01T10:00:00",
                    "descripcion": "Área de producción",
                    "id_empresa": "company-123",
                    "id_proceso": 1,
                    "is_active": true,
                    "nombre": "Producción de químicos",
                    "tipo_proceso": "produccion",
                    "updated_at": "2025-03-01T10:00:00"
                }
            },
            "ProcessUpdateDTO": {
                "properties": {
                    "nombre": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nombre",
                        "description": "Nombre del proceso"
                    },
                    "descripcion": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Descripcion",
                        "description": "Descripción del proceso"
                    },
                    "tipo_proceso": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tipo Proceso",
                        "description": "Tipo de proceso"
                    },
                    "is_active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Active",
                        "description": "Estado del proceso"
                    }
                },
                "type": "object",
                "title": "ProcessUpdateDTO",
                "description": "DTO para actualizar un proceso",
                "example": {
                    "is_active": true,
                    "nombre": "Producción modificada"
                }
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}